Fixing all the Ruby provisioning bugs

+++ scripts/ruby/consul.rb
@@ -10,21 +10,25 @@ STDOUT.puts "Script file: #{__FILE__}"
 FileUtils.mkdir_p "/consul"
 Dir.chdir "/consul"
 
-# Download the consul zip file
-command = "wget '#{ENV['CONSUL_DOWNLOAD_URL']}'"
-`#{command}`
-while $?.exitstatus > 0
-  sleep 5
-  `#{command}`
+# Check to see if consul is already there and skip the download if it is
+consul_exists = `./consul -v` rescue nil
+unless consul_exists
+  # Download the consul zip file
+  command = "wget '#{ENV['CONSUL_DOWNLOAD_URL']}'"
+  output = `#{command}`
+  while $?.exitstatus > 0
+    STDERR.puts output
+    sleep 5
+    output = `#{command}`
+  end
+  # Install zip utilities to unzip the consul binary
+  `apt install -y zip`
+  `unzip -o *.zip`
+  `rm -f *.zip`
 end
 
-# Install zip utilities to unzip the consul binary
-`apt install -y zip`
-`unzip -o *.zip`
-`rm -f *.zip`
-
 # We need the private address for coordination
-self_address = `"$(ip addr show eth0 | grep 'inet ' | awk '{print $2}' | cut -f1 -d'/')"`.strip
+self_address = `ip addr show eth0 | grep 'inet ' | awk '{print $2}' | cut -f1 -d'/'`.strip
 
 # Metadata endpoint for coordination
 metadata_endpoint = "#{ENV['ENDPOINT']}/metadata/consul"
@@ -33,32 +37,37 @@ metadata_endpoint = "#{ENV['ENDPOINT']}/metadata/consul"
 # Do this only if the there are no hosts already
 if `curl -k "#{metadata_endpoint}/hosts.length"`.strip.to_i <= 0
   payload = {hosts: []}.to_json
-  command = "curl -k -XPOST '#{metadata_endpoint}' -d '#{payload}'"
-  `#{command}`
+  command = "set -x; curl -k -XPOST '#{metadata_endpoint}' -d '#{payload}'"
+  output = `#{command}`
   while $?.exitstatus > 0
+    STDERR.puts output
     sleep 2
-    `#{command}`
+    output = `#{command}`
   end
 end
 
 # If we don't have enough registered nodes then loop until we do
-while `curl -k '#{metadata_endpoint}/hosts.length'`.strip.to_i < ENV['NODE_COUNT'].to_i
+while (output = `curl -k '#{metadata_endpoint}/hosts.length'`).strip.to_i < ENV['NODE_COUNT'].to_i
+  STDERR.puts output
   # It is possible some other node reset everything so make sure we re-register
-  if `curl -k '#{metadata_endpoint}'`[self_address].nil?
-    `curl -k -XPOST '#{metadata_endpoint}/hosts' -d '#{self_address}'`
+  if (output = `set -x; curl -k '#{metadata_endpoint}'`)[self_address].nil?
+    STDERR.puts output
+    STDERR.puts `set -x; curl -k -XPOST '#{metadata_endpoint}/hosts' -d '"#{self_address}"'`
   end
   sleep 1
 end
 
 # Whoever registered first will set a key
-if `curl -k '#{metadata_endpoint}/hosts/0'`[self_address]
+if `set -x; curl -k '#{metadata_endpoint}/hosts/0'`[self_address]
   key = `./consul keygen`.strip[0...24]
   payload = {key: key}.to_json
-  `curl -k -XPOST -d '#{payload}' '#{metadata_endpoint}'`
+  STDERR.puts `set -x; curl -k -XPOST -d '#{payload}' '#{metadata_endpoint}'`
 end
 
 # Wait for the key to be set
-while `curl -k '#{metadata_endpoint}.keys'`["key"].nil?
+while (output = `curl -k '#{metadata_endpoint}.keys'`)["key"].nil?
+  STDERR.puts "Waiting for key to be set"
+  STDERR.puts output
   sleep 1
 end
 
@@ -69,7 +78,7 @@ end
 encryption_key = `curl -k '#{metadata_endpoint}/key'`.tr('"', '')
 main_host = `curl -k '#{metadata_endpoint}/hosts/0'`.tr('"', '')
 command = [
-  "nohup ./consul agent -ui -syslog -server -bootstrap-expect #{ENV['NODE_COUNT']}",
+  "./consul agent -ui -syslog -server -bootstrap-expect #{ENV['NODE_COUNT']}",
   "-data-dir '/consul'",
   "-bind '#{self_address}'",
   "-advertise '#{self_address}'",
@@ -77,4 +86,14 @@ command = [
   "-retry-join '#{main_host}'",
   '&'
 ].join(' ')
-`#{command}`
+# Double fork to avoid any weirdness
+Process.fork do
+  Process.setsid
+  p = Process.fork do
+    STDIN.reopen '/dev/null'
+    STDOUT.reopen '/consul/output.log', 'a'
+    STDERR.reopen '/consul/output.log', 'a'
+    exec(command)
+  end
+  Process.detach(p)
+end
